Fix dice and lower pc health

+++ include/dice.hpp
@@ -5,13 +5,13 @@
 
 class Dice {
 private:
-	uint8_t base;
-	uint8_t dice;
-	uint8_t sides;
+	int32_t base;
+	int32_t dice;
+	int32_t sides;
 
 public:
 	Dice() : Dice(0, 0, 0) {}
-	Dice(uint8_t b, uint8_t d, uint8_t s) : base{ b }, dice{ d }, sides{ s } {}
+	Dice(int32_t b, int32_t d, int32_t s) : base{ b }, dice{ d }, sides{ s } {}
 
 	int32_t roll() const {
 		int32_t roll = base;
